Fix (remaining) underlines expanding to the end of the buffer
Summary:
The first diff made every line has its own annotation. So now the only way to get a multi-line is by editing the text.

The ugly behavior is when the line expands all the way to the end of the buffer, the way it is implemented by atom is to set `right: 0`, otherwise it sets `width: x`.

Since modifying the code in a lint warning is likely going to invalidate it anyway, what we can do is to "hide" the regions that would go to the end of the buffer by overriding the inline style `right: 0` with `right: auto !important`.

This is not pretty but should make the lint warnings much less annoying to the users.

Reviewed By: jgebhardt

Differential Revision: D4193838

fbshipit-source-id: 2bd545f7fc73c6286a84a6c6680221314ccf6021