Fix (one case of) underlines expanding to the end of the buffer
Summary:
This looks really terrible when the underline goes all the way to the end of the line. This happens because the method we're using is not supposed to be used for underline but for putting a background.

This diff splits it into several lines so that it has a good behavior when you display the underlines.

Note that if you add a \n in the middle of the line, it's still going to look completely busted with this diff.

Reviewed By: jgebhardt

Differential Revision: D4193607

fbshipit-source-id: 1c5476813bf3818dc64e9c64a7f4e3e3b18f2a58