Fix PHP script launching case when parameters are specified
Summary:
When parameters are specified, only the first token in the launchScriptPath is eligible to be
compared to the path xdebug gives us.  This is hacky and it would be better to have the PHP debugger pass its
params in a separate field of the config, but I'd rather do that at the same time as updating the UI to have a
separate atom input for params and script path.

Reviewed By: yinghuitan

Differential Revision: D4202552

fbshipit-source-id: 040723562016ddcc1fcb3838b868e620e486ff0a