Fix the callstack frame highlighting issue
Summary:
There is a bug that after you stepping/restart debugging callstack window still highlights the old frame. This diff fixes it.
Also, the model we did for the highlighted frame is not aligning with Jonas pattern -- we should not have state in sub-component; every state is pushed as props from master UI.

Reviewed By: johnislarry

Differential Revision: D4205826

fbshipit-source-id: ab993e5c5bdf8a5a6b735f055380608bb876fe1c