Fuzzy filename search: Include name of root directory in index
Summary:
Currently we try to strip off the name of the root directory, but often repositories have prefixes/suffixes that don't quite match what the user entered.

Instead, let's just include the directory in the index so we get fuzzy matching for free.
One downside is that these matches don't get highlighted, but it's rare that this happens so I think it's OK.

Reviewed By: matthewwithanm

Differential Revision: D4209165

fbshipit-source-id: 825a856cff1be380bafd0f07ad22364c10e352e5