Nuclicons
Summary:
Spurred on by facebook/nuclide#890 and vjeux's recent work on this, I decided to take the plunge and create an icon font. Eventually maybe we should devote some time to getting a more robust solution into Atom but this'll work for now.

I also updated some of our components to accept `iconset` props (based on the concept from suda/tool-bar) so that these can be used more generally. The Buck and HHVM task runner task icons are updated and the debugger home fragment.

TODO:

* After switching the Debugger to workspace views, we can connect it to the file-icons service and use the new debugger icon in the tab.
* MOAR ICONS

Reviewed By: vjeux

Differential Revision: D4210136

fbshipit-source-id: 2636df217ddf3fcf6f31747d120e6492bcf02150