Add native-key-bindings class in lower nodes
Summary:
So, native-key-bindings prevents <AtomInput> from receiving backspace and enter events.

Instead of adding the thing to the entire debugger view, I'm adding it to the ListView and LazyNestedValueComponent elements.

I really hope that we don't add any <AtomInput> inside of them... otherwise it's not going to be a fun debugging session, but I think it should be safe.

Reviewed By: johnislarry, matthewwithanm

Differential Revision: D4217440

fbshipit-source-id: b05aed6ca91c509dfb8b4199edbdb111f61f9820