Adapt consumeFirstProvider to syncronous api change
Summary:
https://github.com/atom/service-hub/commit/8a6058ab03352ef600fb958ca0e7fd78fdfb66e5

Made this function synchronous, so we actually get an undefined reference to subscription because the callback is called before consume returns.  I think a good goal would be to get rid of consumeFirstProvider because only the debugger uses it, and I think we can get around needing it by flipping the consumer/provider relationship.

Reviewed By: yinghuitan

Differential Revision: D4217132

fbshipit-source-id: aaba427f20c41f85fae058e71863182792bcfdd5