Handle focus/blur
Summary:
Chrome has an annoying behavior where it triggers a mousemove event after the focus of the application changes (alt tab) or you switch tabs (ctrl tab).

The result is that if you are using the keyboard, when you ctrl-tab, it's going to show the datatip of whatever is below the cursor in that new window. This is never intended :)

The fix is to drop that first mouse move. This is safe even if the behavior changes in the future because it's almost impossible to trigger a single mouse move using the mouse, so dropping one isn't a big deal.

Reviewed By: jgebhardt

Differential Revision: D4210858

fbshipit-source-id: 8430a40cfa5a28a2f8dbc19b1bb26b54d27fe579