Less aggressive closing after interacted with
Summary:
The closing of the datatip was way too aggressive in cases where you want to interact with the datatip.

The two use cases I ran into it are:
- If you want to copy and paste
- If you want to horizontally scroll

In this diff, I add logic to detect those two cases and if the user interacted with it, then debounce the logic to hide if outside. This way it gives you some wiggle room.

For the scrolling, I avoid making the entire view scroll when you are trying to scroll horizontally.

Note: if you press escape or click somewhere else, it's going to instantly hide the tooltip.

Reviewed By: jgebhardt

Differential Revision: D4210575

fbshipit-source-id: a04676b827d212fcf9742b723e287d389f258ba8