Use /DEVELOPMENT file to denote running from source
Summary:
Using the existence of `.flowconfig` to determine if we're running from source is very obtuse. You'd never know that unless you dig deep into the code. This diff created the concept of `DEVELOPMENT` file - when it exists, it means we're running from source, if it doesn't, then we're a build.

Also updated the transpiler to use that instead of detecting the existence of `babel-core`. That fixes #794.

Reviewed By: hansonw

Differential Revision: D4220544

fbshipit-source-id: 2875bd8f083eb3fafe72f4ed70c19bc2296b13c6