Add logging for pinning
Summary:
I want to kill the pin feature of datatips because I don't think it makes sense but I don't have a good sense of how often it is being used. Let's track this and look at the data.

Hopefully almost no one is using it and we can get rid of it :)

Reviewed By: jgebhardt

Differential Revision: D4210905

fbshipit-source-id: 8ff15819808a9decb3a413259cc90baa029ccba1