Show highlight when right clicking on an element
Summary: This has annoyed me for a long time and I'm unsure if Atom people are going to accept my fix, so I want to monkey patch it inside of Nuclide for now. I think that this is going to be a big boost in terms of quality for Nuclide.

Reviewed By: zertosh

Differential Revision: D4227051

fbshipit-source-id: 1c692dfcc9ce6b71916b3df3d3cc0ce578d322f3