Move to Workspace View
Summary:
I used OutlineView as an example and copy pasted as many things as needed.

Note that the architecture is now pretty weird with a ContextViewPanelState that calls into a ContextViewManager just to finally display the ContextViewPanel. Ideally, we should be able to remove some of those layers of abstractions but this is the minimal amount of work needed to make it a workspace view.

Reviewed By: matthewwithanm

Differential Revision: D4235169

fbshipit-source-id: 7d9850ef0fe574e889f6c1a037ea795b44381602