Don't hide on blur
Summary: This fixes a small edge case but makes the entire tooltip not interactive anymore as it causes a blur when you click on anything inside of the tooltip. I haven't seen any way to figure out if the blur is caused by a click on the tooltip so let's just remove it.

Reviewed By: nmote

Differential Revision: D4244676

fbshipit-source-id: cff77777c939bea2fd160837fa06888006af6c47