Task Runner: Display a "Select Task" dropdown in the null state
Summary:
Now that the task runner can get into a null state if a task becomes disabled, we should change the null state to indicate that tasks are available via the dropdown plus a placeholder "Select Task" label.

Furthermore, disable the dropdown altogether if all of the tasks are disabled.

Reviewed By: matthewwithanm

Differential Revision: D4242403

fbshipit-source-id: 567971483710e736ba6aba8233f35fe30ec12470