Fix hphpd warning notification
Summary:
If the user has an hphpd session attached we are supposed to display a notification warning them.  This notification is delivered from the server to the client via an observable, but the client was never receiving the message.  It turns out this is because of a race.  Chronologically the events look like:

1. Client waits for the server to initialize the PHP debugger service
2. PHP debugger service sees that there's an hphpd session, and sends the warning through the observable channel
3. PHP debugger service finishes initializing, and client continues execution
4. Client subscribes to the observable streams

Since the client subscribes after the message is sent, we don't display the notification.  I fixed this by using a ReplaySubject, which re-issues all the elements its seen when new subscribers come.

Reviewed By: yinghuitan

Differential Revision: D4263167

fbshipit-source-id: c1400a48e231a513e11cd1bf1bcbe1a0975e0463