Create a core:loaded-shell-environment hook package
Summary:
The current process.js code listens for the `core:loaded-shell-environment` hook, but this essentially relies on the fact that process.js loads before the hook triggers. There's always the chance that we miss it.

Instead, create a nuclide-loaded-shell-hook package whose sole purpose is to listen for the activation hook, and triggers `commons-node/loadedProcessEnvironment()`.

Reviewed By: vjeux

Differential Revision: D4274408

fbshipit-source-id: bcc3ab223df2e0d437b374f6edf0f57f1d1a8fdd