Show debugger in workspace view; remove legacy debugger view
Summary: The purpose of this commit is to move the debugger into workspace views. I intentionally tried not to change any behavior, however, access to the legacy debugger view needs to be removed in order for this to work. That's because workspace views are removed from the DOM when dragging/moving them, but removing the Chrome UI isn't allowed (the webview would lose its state).

Reviewed By: yinghuitan

Differential Revision: D4209242

fbshipit-source-id: 96b9800eaa40d432505d8f76bda80ba7ac8fcc05