Restore hide on blur functionality
Summary:
In D4210858 I introduced a fix to ensure that tooltips do not show up when you switch app or tabs and do not move the mouse.

Unfortunately, it's been reverted in D4245752 and D4244676 because it made it impossible to pin a datatip.

This diff brings back the behavior and fixes the tooltip pinning. The fix works by making sure that if you have your mouse inside of the datatip, we don't hide it on blur or focus.

Reviewed By: jgebhardt

Differential Revision: D4275188

fbshipit-source-id: 5924b7def1ad0cc19c8abb4839187b630c074279