Pretty-print flow types
Summary:
In D4012232, I introduced a pretty-printer for flow but the integration was a bit awkward. But waleedk in D4263691 came up with a very clever solution. Always pretty-print the ocaml type but truncate it to 100 characters and expand on click.

I think this would work very well for flow as well!

cc avikchaudhuri, jeffmo, I would love for flow to pretty-print the types for us!

Reviewed By: nmote

Differential Revision: D4272628

fbshipit-source-id: 192feae2389400f06cbda6c88e0684d429816b41