Remove iconset props, always prepend "icon-"
Summary:
Given D4293334, I think we might as well just assume that icon classes are always prefixed with "icon-". The Atom team floated the idea of adding a `getIconClassName()` method that didn't prefix, but honestly prefixing isn't a big deal.

In that case, we can roll back the "iconset" additions and just always prepend "icon." This means we can't use our utilities with icon fonts that don't prepend "icon-" to the classnames (ion, etc.) but 1) we aren't doing that and 2) we probably shouldn't because we don't have real guarantees that they'll be loaded.

As an added benefit, this made it easy to add back the type checking; Nuclicon names are now enumerated too and I updated the things that were using iconset/icon string pairs to accept a single `atom$Octicon | Nuclicon` icon instead.

Reviewed By: vjeux

Differential Revision: D4297725

fbshipit-source-id: 8d530e52f74d045912408fc21255885de1876bfa