Make workspace-views API more Atom-like
Summary:
Continuing the effort to get to an API that can be merged into Atom core, this commit updates the `open()` API to be more like Atom's `workspace.open()` and the Location API to more closely mirror Atom's Panes'. This has the side-effect of enabling items to be added without being focused.

In addition, I added a new action for when the initial packages have been activated, allowing for clean testing.

Reviewed By: shushz

Differential Revision: D4296720

fbshipit-source-id: b4e724726491b34da35660cd9231d216aec6c401