Diff View: Fetching file changes more efficiently and accurately
Summary:
Previously, I was caching the results of the command: `$ hg log --rev $rev --template {files}`
for every commit in the stack, then, locally merge them approximately to come to the stack changes.
While this is cachable, it doesn't scale for big repos since this command requires accessing an expensive index.

Also, the ending result isn't accurate - you could have reverted a file, and it still shows as modified in the comparison as the file contents aren't compared.

Instead, here I use `hg status --rev $compareId` which is pretty fast (3 seconds) even for a 10 commits stack on a large repository.

Reviewed By: vjeux

Differential Revision: D4289634

fbshipit-source-id: f74c7a6a62653f4fc76764c75ae238a074e87276