Do a second query for revision info
Summary:
Right now, the way blame works is that we are doing hg blame and use the json output to display what's in the gutter.

This is a bit unfortunate because the blame output is limited in what informations it can give us, for example we cannot get the commit title, which is extremely important as otherwise you've got no idea what the commit is about.

Instead of creating another datastructure for what a commit is and doing parsing, we can query the list of revisions in a second step.

This is also what quark-zju told me would be the fastest way to go about it once we switch to his fastannotation extension.

Reviewed By: matthewwithanm

Differential Revision: D4266029

fbshipit-source-id: 95c7d0fe5956e7651db7fa4b0d90a29ebdcbc07a