Upgrade to ESLint 3.12.0
Summary:
My PR for no-await-in-loop got merged: https://github.com/eslint/eslint/pull/7563

That was the last thing we needed eslint-plugin-babel for, so I've dropped the
dependency on that.

This diff contains the changes necessary due to the lint rule name change (now
it's called just `no-await-in-loop` rather than `babel/no-await-in-loop`, so ignore comments needed updating).

This diff also fixes some lint errors caused by changes to the `indent` lint
rule.

Reviewed By: zertosh

Differential Revision: D4308781

fbshipit-source-id: 2cd4584205fde13127e7f84ee79504b6b2333aa3