Dropdown: add submenu option type
Summary:
Allow option type of 'submenu' in Dropdown, which maps to 'submenu' in Electron.

This is a basic implementation, we could do more UX-wise in case we start using the component in more places, like showing breadcrumb buttons when a nested option is selected.

The dropdown button label when a nested option is selected is the label of the final option. If you want to display e.g. anything related to the parent, you need to define it manually in the option via selectedLabel.

Reviewed By: matthewwithanm

Differential Revision: D4304570

fbshipit-source-id: bb572efa50bb63cbb0911771f59017875da67437