Do not unmount React component on update
Summary:
Problem 1: Calling `ReactDOM.render()` on an already rendered component applies the React diffing algorithm. But, if you call unmountComponentAtNode it is going to force remove the component and remount it from scratch, which will do a lot of DOM and React operations for no good reason.

Problem 2: The second issue is that we have a React wrapper for a custom element that mounts when it gets attached to the dom and unmounts when it gets removes, coupled with the fact that atom has an off-screen "buffer" for decorations. So, everytime it gets moved between the two buffers it unmounts and remounts the React component!

The fix for that is to manage the lifecycle ourself. Whenever we create the marker we render, and whenever the marker is destroyed we unmount.

Problem 3: For some reason the code hardcodes the width of the window and re-renders when it changes. There must have been some good reason for doing that but it seems like the diff view and comments work well without, so I'm not really sure what's going on, but I know that it's going to cause useless re-renders.

Reviewed By: mostafaeweda

Differential Revision: D4311536

fbshipit-source-id: 00c8c64c8e4ac2584951f2aeb34db43cce985b30