Make it awesome
Summary:
Each row is separate so we don't have a lot of flexibility. After lots of trials and errors, came up with a profile picture and a vertical line that ends with a dot, to remind of the interactive smartlog design.

The background of the entire gutter is using the same stripes as the diff view. This way, we don't need to add a gutter for every single line, even when the user interacts with it.

Added a tooltip to show the full title, date and author unix name.

The right bar fades as the revision is older using a simple interpolation where most recent diff is 1 and least recent is 0.2.

When collapsed, all the rows are stacked under each other but because there's a solid background, you can only see a single one. This is not technically correct but at least it doesn't look broken.

Reviewed By: jgebhardt

Differential Revision: D4271964

fbshipit-source-id: 8040a70b7a46929a2b35d4480debe1b25abcfbd4