Hide if mouse leaves the textarea
Summary:
There's a very annoying behavior currently when: you move your mouse outside of the texteditor but the last thing you cross happens to have a tooltip.

After we finish fetching the tooltip, we look if the last known mouse position is above the word, which unfortunately it is because we only update it on mousemove within the textarea.

The solution is easy, if we move outside of the area, we want to clear the last move event and hide the tooltip if already there.

Reviewed By: jgebhardt

Differential Revision: D4313403

fbshipit-source-id: 1c0e455d501a254308cac902cc6c95e9d0732be8