Switch to generating flow_project_1 fixture
Summary: I'm trying to centralize the path rules for transpiling, and this fixture is giving me grief. Transpiling it breaks the tests because the line numbers change. To workaround that, I'm switching to generating he fixtures on-the-fly. This actually makes more sense for flow fixtures since that way we don't have an actual `.flowconfig` on disk (so no accidental flow roots).

Reviewed By: nmote

Differential Revision: D4315127

fbshipit-source-id: f62fd2d1c7bc50a90ff25847c788ceef8b24e807