Port nuclide-node-transpiler tests to jasmine
Summary: The tests for `nuclide-node-transpiler` where plain js files with `assert`s. `npm test` ran each file with `find spec -name '*-spec.js' -print0 | xargs -n 1 -0 node` so if any `assert` failed the command would return some non-zero value. Since `nuclide-jasmine` used `nuclide-node-transpiler`, it couldn't be used as the test runner. This whole setup is weird since nothing else does this. This diff ports the tests to using plain `jasmine-node` so there's some consistency with the rest of Nuclide. It's not quite the same as `nuclide-jasmine` because it doesn't have the focus helpers or transpiling, but it's close enough.

Reviewed By: hansonw

Differential Revision: D4323892

fbshipit-source-id: fb60260e055ab6745f7e376c9c88f6d9fb2bdd5f