Allow providers to register platform groups in buck plugin
Summary: Add support for platform groups instead of only single platforms. Devices are displayed in a submenu for each platform. If the provider only sends a single platform, everything is shifted "one level up", so devices are listed directly in the dropdown group, which is better UX (and also the current behavior).

Reviewed By: hansonw

Differential Revision: D4314884

fbshipit-source-id: e25e8b60b11bc5b3e873dc802dc522fd4a0fc90e