Give registered provider function access to platform service
Summary:
Pass platform service to provider functions that consuming plugins register. This way we can e.g. pass the existing buck service to the registered providers, so they can utilize it if they need to.

This also changes ownership of the platform service from a global variable in main to BuckBuildSystem.

Reviewed By: hansonw

Differential Revision: D4315183

fbshipit-source-id: e47bd11ea8047456daea130dd5e641b38dbe4aab