Clang: easy server-side cache optimization
Summary:
The server-side cache uses binary search to filter the list of candidates to avoid expensive C-Python boundary calls, especially when the result set is > 100.

An easy optimization to make is to progressively narrow the binary search window if the given prefix starts with the previous one.

Empirically, the 2nd and 3rd keystrokes when typing `abc` are about 50% faster (they're already < 20ms, but this brings them to < 10ms).

Reviewed By: nmote

Differential Revision: D4334267

fbshipit-source-id: d3853a165813aca5602f0516d4133a28c893b32e