Add warning when rename shortcut fails
Summary:
I kind of wanted to reuse the UI in the modal that lets people know when there are no refactorings, but this has two advantages:

1) It's simple to implement
2) It is ephemeral like other similar notifications. This is especially important if you are in the middle of something and just hit the hotkey to rename.

The main disadvantage is that the "waiting for refactorings" modal flickers before this appears. It might be worth waiting a couple hundred milliseconds before displaying that (both for the rename shortcut and the more general refactoring UI).

Reviewed By: hansonw

Differential Revision: D4334412

fbshipit-source-id: 0c295d248954d25efc20388fe57c541d3de96621