Improve breakpoint experience for mobile js debugging
Summary:
This refactors the line based breakpoint-related functionality into its own class.  The diff after this does exception breakpoints.  It also handles the case when there are no connections correctly.  This used to cause a bug where the user would:

1. Set a bp
2. Attach the debugger
3. The bp would be removed

Because Nuclide would request to set a bp after attaching, but there were no connections so this request was answered with a generic error response.  Now we wait to respond until there is a connection, so we don't drop breakpoints anymore.

yinghuitan  I thought about what you were saying about waiting to resume the loader bp until the first connection comes in, but that wouldn't address this problem if later on all the jscontexts end, the user sets a bp, and then a new jscontext is spun up.  I think this way is more general and would handle that situation.

Reviewed By: yinghuitan

Differential Revision: D4310217

fbshipit-source-id: 4a31670ccbf53a6760e07068dfad0dc81073269a