Task Runner: Remember visibility state based on active project root instead of Atom's project
Summary:
The reconstruction of the task runner's visibility state, like most of Nuclide's deserialization, has been based on Atom's serialized package state (which is done per-Atom project at startup). Unfortunately, what we think of as projects aren't necessarily available at startup?especially in the case of remote directories, which are mounted into an empty Atom "project" later.

With this commit, we step outside of Atom's serialization and do our own per-project-root. Truthfully, a lot more of Nuclide's state (including other aspects of the task runner) should probably be serialized like this. It's probably worth figuring out a solution that'll be easy to implement across packages and with less boilerplate.

A bunch of this commit is just shuffling a lot of the logic that was in the reducer (and corresponding tests) into epics. This allows us to keep localStorage access out of the reducer.

Reviewed By: shushz

Differential Revision: D4336234

fbshipit-source-id: ef5e1bb69ca7794ee7dc18561f1d0e6db1e95345