Make nuclide-ocaml's hyperclick provider less greedy
Summary:
Before, `nuclide-ocaml`'s hyperclick provider always returned a match even if it would later turn out that merlin had no results. This had the unfortunate effect of crowding out lower-priority providers (e.g. hyperclicking on a url in a comment didn't work). This diff fixes that so that we only return a match if there are results.

Also, error notifications are only displayed for `pushNewBuffer` errors. When there is no result, no error is shown. This matches the behavior of all other Nuclide hyperclick providers.

Reviewed By: hansonw

Differential Revision: D4346667

fbshipit-source-id: d69d031411d45e3a1c1a52698b22583804361130