Remove url.parse() from nuclideUri common flows
Summary:
Node's `url.parse` is a poor tool for parsing Nuclide URIs. It's horribly slow, it has a ton of corner cases, which required several ad-hoc workarounds and it generally just wasn't meant to be used for paths.
In addition to the above, the structure of Nuclide URIs is extremely simple, it is just not justified to use such a complex machinery to parse that.

This diff removes the `url.parse()` usage from `nuclideUri.parse()` in favor of an extremely simple string matching and splitting logic.

Reviewed By: peterhal

Differential Revision: D4335604

fbshipit-source-id: 5233989e3684bf0ce55c255256f8793c46ac3bf7