Make it group on remote connections
Summary:
I was doing a === comparison but it is getting serialized. Fortunately the hash is unique so we can use that.

A bigger and better fix would be to send an array of ids and then an array of values to avoid sending the same thing 10 times, but let's make this one now that fixes the bug.

Reviewed By: mostafaeweda

Differential Revision: D4362894

fbshipit-source-id: 9ba07ab5573e3389caebddfa2c2fe8b73d5d46b8