Remove special build entry file magic in favor of __DEV__ checks
Summary:
Instead of swapping the `main` from the Nuclide client package at build time, add a `__DEV__` check before loading the transpiler. The idea behind the `main` swap was to avoid loading the transpiler in prod builds, but this is weird behavior. Now if you try to load the transpiler in a prod build, it'll throw.

The next step is to remove `const {__DEV__} = require('../../nuclide-node-transpiler/lib/env');` altogether in favor of a transform. But this will have to be a special post-build step because most of the files that have it are non-transpiled files.

Reviewed By: hansonw

Differential Revision: D4364549

fbshipit-source-id: cee719e1b267207937a6d96891bdc58036155498