Add control over "Open All" to quick-open providers
Summary:
This diff introduces the idea of "canOpenAll" to quick-open providers. They can now decide whether to allow quick-open to open all of the results or not. This is useful if you know that you have a lot of results or your files are really expensive for some reason.

The default is to allow "open all", which was the previous behavior. Except for OmniSearch - you can't "open all" for OmniSearch anymore, since you can't ask the individual providers if it's ok to open everything. Besides, OmniSearch has too much random stuff, doubt anyone was using this.

Also fixed some of the types. Provider specs do not have optional properties, when they're "baked", anything that's missing is filled in with a default.

Reviewed By: jgebhardt

Differential Revision: D4371844

fbshipit-source-id: a191f89feef82094a6d4d581df0a4f8b649dd7d3