De-singleton quick-open
Summary: Converts `quick-open`'s classes from singletons to instances. After this change, it is possible to create a new set of "quick-open" classes whenever it's opened. This allows us to clean up a lot of the existing code because don't have to worry about expiring caches, and we don't have to worry about updating state when Atom's project state changes (since you can't really add a project while "quick-open" is open, and closing-opening it reads a fresh state). This is going to simplify adding more functionality.

Reviewed By: jgebhardt

Differential Revision: D4373697

fbshipit-source-id: fdc4f81b397ca73d1e197aebe3536f30a8327f4d