Use values for quick-open provider config
Summary:
This dramatically simplifies the quick-open provider interface by switching to primitive values for things that are constant (like `name` and `title`) instead of functions. Also cleaned up the types so the provider type is the key for a tagged union.

The most notable difference in the interface is that the `isRenderable` has been replaced by a `display` object. The presence of a `display` object implies that this provider is renderable, and makes the distinction between required and optional properties for rendering much clearer.

After this change there is no real need to have the concept of a "baked provider" (or `ProviderSpec`), so I'll be removing that soon.

Reviewed By: jgebhardt

Differential Revision: D4387568

fbshipit-source-id: 1c077e0eef35eae26d9499e72bef75abc59371b9