Use last results on initial quick-open render
Summary:
This fixes an issue where you'd see a flash of "Search away..." or "No results..." under these circumstances:

* When you closed, and re-opened quick-open without changing the query.
* When you switched to another provider and back without changing the query.

Now, whenever there's a provider change, instead of showing empty results, we show whatever results we have in the cache for that query. Since we still fire off a re-query, if anything did in fact change, you'd see the UI update.

Reviewed By: aadisriram

Differential Revision: D4391282

fbshipit-source-id: 18a9edd1b75637f32f7f420c331e7356862a84d3