Fix quick-open selection flickering
Summary:
This fixes an issue where the result selection would flicker as you typed or if you switched providers.

Before, whenever a provider switch or search update started, the selection was removed, then it was added back to the top result when the update finished. This caused the flickering. Now, we never remove the selection. Instead, we move the selection only when the user explicitly takes an action that should move it.

Reviewed By: jgebhardt

Differential Revision: D4391286

fbshipit-source-id: 5893b9c89cd167b32cd54315428a66cf2583e1a3