Use flexbox to size the test runner component
Summary:
Facebok:
Previously, we were hardcoding the height of the test runner component, wasting tons of space:
https://pxl.cl/6wrk

This is a hotfix to make it so that we fill the available space:
https://pxl.cl/6wrm

TBH I'm not super happy with this solution, but I don't own this code (and I'm not sure who does...) and just wanted to quickly fix the problem.
That being said, the Test Runner needs some serious UI love.

Reviewed By: zertosh

Differential Revision: D4395243

fbshipit-source-id: a9479e5a4a0f8996e5704bd2dc2fe6722a9af375