Support starting the nuclide debugger with init scripts
Summary:
Currently, there are two ways of starting the nuclide native debugger: with a
pid to attach to, or with an executable to launch. For more complex debugging
scenarios (e.g.: connecting to a remote debug server running on a phone), we
need more flexibility, and we need to be able to start the debugger with an
init file containing a set of commands to run.

Original version of this patch by Walter Erquinigo <wallace@fb.com>.

Reviewed By: yinghuitan, a20012251

Differential Revision: D4364852

fbshipit-source-id: 0ce1fd5d4582f4344eabeba918b419eb13cdda13