Add templates for Nuclide
Summary:
While writing this diff, I tried to create a template for every single type of file I encountered. This was a very valuable exercise :)

Please feel free to nit away in those templates, I'm not super confident in them being the source of truth.

Side note, It's weird that the less files do not have the fb license headers.

Reviewed By: vjeux

Differential Revision: D4283356

fbshipit-source-id: a4aac6ca219dd2dd192f20d93cdab587d8ffd4c4