Polish Commit Workflow 6/n: Abort publish if a needed amend failed
Summary:
This diff standardizes the console as the target for all process output, regardless of the flow that started it - instead of losing that valuable stdout/stderr somewhere in the process.

Also, aborting the chain on failed amend (exit code != 0) will gurantee no cascaded publish failures are encountered (when dirty changes) or non-updated commit messages are diffed + give the user the immediate sense of progress.

Reviewed By: aadisriram

Differential Revision: D4415892

fbshipit-source-id: c3c732de94e4b6632129c42bd2eafa78a9e95d23