Add remote regexp find/replace implementation in GrepService
Summary:
Implement a remote function that performs search/replace on a set of remote paths and streams back results.

This is meant to roughly match the `scanhandler` implementation at https://github.com/atom/scandal/blob/master/src/path-replacer.coffee

Note that this works on a line-by-line basis and as-such, does not support multiline regexps. Atom's code does this as well, so it's probably best that we're consistent here.

Used in the next diff.

Reviewed By: zertosh

Differential Revision: D4418537

fbshipit-source-id: f8e4f56056664c0d9dbedb852080874343abf843