Monkey-patch atom.workspace.replace to work remotely
Summary:
This is kinda gross, but replace `atom.workspace.replace` with a custom implementation that works for remote files.

We leave open files and local files to Atom to minimize the differences.

(Ideally `atom.workspace.replace` would shell out to `Directory.replace` and our `RemoteDirectory` implementation could override this... but that's probably some ways off).

Reviewed By: zertosh

Differential Revision: D4418809

fbshipit-source-id: 78197fe8b81f4b2c5339e4df4cf97032c453ad3d