Rename JS debuggers to be more accurate
Summary:
`iOS Webkit Debug Proxy -> Mobile JS`.  Much clearer that the intention is to debug javascript running on a mobile device.
`JavaScript -> NodeJS`.  Much clearer that this is only for debugging node programs, as opposed to JS running in a browser or on a phone.

Changelog: Clearer names given to JavaScript debuggers.

Differential Revision: D4430502

fbshipit-source-id: 1c72afc2cb9ecab7f20e21ffe8e8b519d8f444f0