Create new QuickSelectionComponent instances on each quick-open invocation
Summary:
Before, the instance of `QuickSelectionComponent` was reused for every quick-open invocation until quick-open was deactivated. After lots of refactoring, we're now able to create new instances of `QuickSelectionComponent`  whenever QO is invoked. Although this sounds like a lot of work, there's no noticeable perf penalty to it. However, there's a big perf win outside of QO, because the event handles on the document and the body are disposed when QO is closed. This means we're no longer doing random work even though QO is open.

This also made it possible to keep the query input always focused. So when you switch providers with the mouse, you don't see the focus ring disappear and reappear.

Reviewed By: jgebhardt

Differential Revision: D4419013

fbshipit-source-id: 7b74c3ccab4c663174999bed20a94c6d207bda42