Change ScopeUpdate to update all scopes at once
Summary:
In favor of D4415889.

This implements our `ScopeUpdate` action to do the Right Thing which is to send all updated scopes at once.  Before, this action was an append operation, which is more complicated to handle correctly.  This change will fix a bug where console evaluation caused the scopes to duplicate in the scopes window.

Differential Revision: D4424709

fbshipit-source-id: 96a638be1c257e42f0fe93bdf3b7750fa8a0d5fd