null return prevents caching in AutocompleteCacher
Summary:
When I wrote this code, I made the assumption that we could tell whether to
cache and filter or not based only on the requests. However, in certain cases we
want the ability for a provider to indicate later that it does not want to
provide any results without forgoing the opportunity to provide results later
for the same token.

This capability will simplify the addition of AutocompleteCacher to the language
service framework by allowing us to put providers completely behind the
AutocompleteCacher while still giving them this flexibility.

Note that in order to maintain performance we have to send out speculative
requests, as explained in the comments. That is why the tests needed to be
updated.

Reviewed By: peterhal

Differential Revision: D4425828

fbshipit-source-id: f01cdc1bfef69e7a4e34f6e4889998095d8c2c86